+++ lib/yaks/serializer/hal.rb
@@ -5,7 +5,13 @@ module Yaks
     class Hal < self
       Serializer.register self, :hal, 'application/hal+json'
 
+      DEFAULT_OPTIONS = { singular_links: [:self, :profile] }
+
       protected
+
+      def options
+        DEFAULT_OPTIONS.merge super
+      end
 
       def serialize_resource(resource)
         # The HAL spec doesn't say explicitly how to deal missing values,
@@ -25,7 +31,7 @@ module Yaks
 
       def serialize_link(memo, link)
         hal_link = {href: link.uri}
-        hal_link.merge!(link.options.reject{|k,_| k==:templated})
+        hal_link.merge!(link.options.reject{|k,_| k.equal?(:templated)})
         hal_link.merge!(templated: true) if link.templated?
 
         memo[link.rel] = if singular?(link.rel)