+++ spec/acceptance/json_shared_examples.rb
@@ -1,0 +1,8 @@
+RSpec.shared_examples_for 'JSON output format' do |yaks, format, name|
+  let(:input)  { load_yaml_fixture(name) }
+  let(:output) { load_json_fixture("#{name}.#{format}") }
+
+  subject { yaks.serialize(input) }
+
+  it { should deep_eql output }
+end